Break up actions and reducers into separate files
These are getting large enough that we want to split them up. This
should do that without changing any functionality.
