+++ web/app/actions/router.ts
@@ -1,0 +1,22 @@
+// Copyright:: Copyright (c) 2016 Chef Software, Inc.
+//
+// The terms of the Evaluation Agreement (Bldr) between Chef Software Inc. and the party accessing
+// this file ("Licensee") apply to Licensee's use of the Software until such time that the Software
+// is made available under an open source license such as the Apache 2.0 License.
+
+export const ROUTE_CHANGE = "ROUTE_CHANGE";
+export const ROUTE_REQUESTED = "ROUTE_REQUESTED";
+
+export function routeChange(newRoute) {
+    return {
+        type: ROUTE_CHANGE,
+        payload: newRoute,
+    };
+}
+
+export function requestRoute(requestedRoute: Array<any>) {
+    return {
+        type: ROUTE_REQUESTED,
+        payload: requestedRoute
+    };
+}
