+++ web/app/explore-page/_explore-page.scss
@@ -1,26 +1,27 @@
 .bldr-explore {
-    li a {
-        @include span-columns(3);
-        @include pad($small-spacing);
-        margin: $base-spacing;
-        border: 1px solid;
-        display: block;
-        
-        .title {
-            display: block;
-            font-size: 120%;
-        }
-        
-        .info {
-            text-align: right;
-        }
-        
-        .stars {
-            font-size: 75%;
-        }
-        
-        .stars:after {
-            content:  "★"
-         }
+  li a {
+    @include span-columns(3);
+    @include pad($small-spacing);
+
+    margin: $base-spacing;
+    border: 1px solid;
+    display: block;
+
+    .title {
+      display: block;
+      font-size: 120%;
     }
+
+    .info {
+      text-align: right;
+    }
+
+    .stars {
+      font-size: 75%;
+    }
+
+    .stars:after {
+      content:  "★"
+    }
+  }
 }
