+++ web/app/reducers/packages.test.ts
@@ -1,18 +1,18 @@
 import {Record} from "immutable";
-import {rootReducer} from "./rootReducer";
-import {SET_CURRENT_PACKAGE} from "./actions/index";
+import packagesReducer from "./packages";
+import {SET_CURRENT_PACKAGE} from "../actions/index";
 
 describe("rootReducer", () => {
-    describe("SET_CURRENT_PACKAGE", () => {
+    describe(SET_CURRENT_PACKAGE, () => {
         describe("when the package does not exist", () => {
             it("sets currentPackage to null", () => {
                 const state = Record({
-                    currentPackage: undefined, packages: [{ name: "nottest" }]
+                    current: undefined, all: [{ name: "nottest" }]
                 })();
                 const action = {
                     type: SET_CURRENT_PACKAGE, payload: { name: "test" }
                 };
-                expect(rootReducer(state, action).get("currentPackage"))
+                expect(packagesReducer(state, action).get("current"))
                     .to.equal(null);
             });
         });
@@ -20,12 +20,12 @@ describe("rootReducer", () => {
         describe("when a package has no dependencies key", () => {
             it("sets it to an empty array", () => {
                 const state = Record({
-                    currentPackage: undefined, packages: [{ name: "test" }]
+                    current: undefined, all: [{ name: "test" }]
                 })();
                 const action = {
                     type: SET_CURRENT_PACKAGE, payload: { name: "test" }
                 };
-                expect(rootReducer(state, action).get("currentPackage")
+                expect(packagesReducer(state, action).get("current")
                     .dependencies).to.deep.equal([]);
             });
         });
@@ -33,12 +33,12 @@ describe("rootReducer", () => {
         describe("when a package has no buildDependencies key", () => {
             it("sets it to an empty array", () => {
                 const state = Record({
-                    currentPackage: undefined, packages: [{ name: "test" }]
+                    current: undefined, all: [{ name: "test" }]
                 })();
                 const action = {
                     type: SET_CURRENT_PACKAGE, payload: { name: "test" }
                 };
-                expect(rootReducer(state, action).get("currentPackage")
+                expect(packagesReducer(state, action).get("current")
                     .buildDependencies).to.deep.equal([]);
             });
         });
